Initial code checkin for EPS library
This initial checkin is complete for Maven and GitHub dependencies and is in the correct structure for both, but is missing its dependency on the GraphicsToolkit as the POM file gives an error in Eclipse until I hook Eclipse up to this GitHub project vs. its local EpsToolkit project. So another checkin will probably follow this one, as it seems unlikely that the current POM file will work on someone else's system unless they have already downloaded a copy of the GraphicsToolkit, which is used by this library.