Update pom.xml
FOund the correct syntax for adding a dependency to the POM for the GraphicsToolkit, but actually it was correct on all previous tries as well. There is a horrible bug in Eclipse 2020-06 with its Maven plug-in and some stick-bit never gets unstuck even after replacing the POM with another one and editing the differences. Eclipse or its Maven plug-in, keep thinking there are spaces and invalid characters in the file and thus produce bogus error messages. This is going to be hard to work around, and means all POM editing will need to take place using macOS's TextEdit utility outside of Eclipse.